+++ app/src/main/java/com/sysdata/demo/accordionview/SampleExpandedViewHolder.java
@@ -22,9 +22,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import com.sysdata.widget.accordionview.ExpandableItemHolder;
-import com.sysdata.widget.accordionview.ExpandedViewHolder;
-import com.sysdata.widget.accordionview.ItemAdapter;
+import com.sysdata.widget.accordion.ExpandableItemHolder;
+import com.sysdata.widget.accordion.ExpandedViewHolder;
+import com.sysdata.widget.accordion.ItemAdapter;
 
 /**
  * Created on 06/04/17.
