+++ app/src/main/res/layout/activity_sample.xml
@@ -23,7 +23,7 @@
     android:background="#E6BDBDBD"
     tools:context="com.sysdata.demo.accordionview.SampleActivity">
 
-    <com.sysdata.widget.accordionview.FancyAccordionView
+    <com.sysdata.widget.accordion.FancyAccordionView
         android:id="@+id/alarms_recycler_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
