+++ fancyaccordionview/src/androidTest/java/com/sysdata/demo/accordionview/ExampleInstrumentedTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.sysdata.widget.accordionview;
+package com.sysdata.widget.accordion;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
