+++ fancyaccordionview/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
   ~ limitations under the License.
   -->
 
-<manifest package="com.sysdata.widget.accordionview"
+<manifest package="com.sysdata.widget.accordion"
           xmlns:android="http://schemas.android.com/apk/res/android"
     >
