+++ fancyaccordionview/src/main/java/com/sysdata/widget/accordion/AccordionLinearLayoutManager.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.sysdata.widget.accordionview;
+package com.sysdata.widget.accordion;
 
 import android.content.Context;
 import android.support.v7.widget.LinearLayoutManager;
