+++ fancyaccordionview/src/main/java/com/sysdata/widget/accordion/CollapsedViewHolder.java
@@ -14,50 +14,37 @@
  * limitations under the License.
  */
 
-package com.sysdata.widget.accordionview;
+package com.sysdata.widget.accordion;
 
 import android.animation.Animator;
 import android.animation.AnimatorListenerAdapter;
 import android.animation.AnimatorSet;
 import android.animation.ObjectAnimator;
-import android.animation.PropertyValuesHolder;
-import android.content.Context;
 import android.graphics.Rect;
-import android.graphics.drawable.Drawable;
-import android.graphics.drawable.LayerDrawable;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.widget.RecyclerView.ViewHolder;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.sysdata.widget.accordionview.utils.AnimatorUtils;
-import com.sysdata.widget.accordionview.utils.SystemUtils;
-import com.sysdata.widget.accordionview.utils.ThemeUtils;
+import com.sysdata.widget.accordion.utils.AnimatorUtils;
+import com.sysdata.widget.accordion.utils.SystemUtils;
 
 import java.util.List;
 
-import static android.view.View.TRANSLATION_Y;
-
 /**
- * A ViewHolder containing views for an item in expanded state.
+ * A ViewHolder containing views for an item in collapsed stated.
  */
-public abstract class ExpandedViewHolder extends ArrowItemViewHolder {
+public abstract class CollapsedViewHolder extends ArrowItemViewHolder {
 
-    protected ExpandedViewHolder(View itemView) {
+    protected CollapsedViewHolder(View itemView) {
         super(itemView);
 
-        final Context context = itemView.getContext();
-        itemView.setBackground(new LayerDrawable(new Drawable[]{
-                ContextCompat.getDrawable(context, R.drawable.alarm_background_expanded),
-                ThemeUtils.resolveDrawable(context, R.attr.selectableItemBackground)
-        }));
-
-        // Collapse handler
+        // Expand handler
         itemView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                getItemHolder().collapse();
-                notifyItemClicked(ItemAdapter.OnItemClickedListener.ACTION_ID_EXPANDED_VIEW);
+                getItemHolder().expand();
+                notifyItemClicked(ItemAdapter.OnItemClickedListener.ACTION_ID_COLLAPSED_VIEW);
             }
         });
 
@@ -65,14 +52,14 @@ public abstract class ExpandedViewHolder extends ArrowItemViewHolder {
             arrow.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
-                    getItemHolder().collapse();
-                    notifyItemClicked(ItemAdapter.OnItemClickedListener.ACTION_ID_EXPANDED_VIEW);
+                    getItemHolder().expand();
+                    notifyItemClicked(ItemAdapter.OnItemClickedListener.ACTION_ID_COLLAPSED_VIEW);
                 }
             });
 
             // Override arrow drawable if running Lollipop
             if (SystemUtils.isLMR1OrLater()) {
-                arrow.setImageDrawable(ContextCompat.getDrawable(itemView.getContext(), R.drawable.ic_caret_up_animation));
+                arrow.setImageDrawable(ContextCompat.getDrawable(itemView.getContext(), R.drawable.ic_caret_down_animation));
             }
         }
 
@@ -82,38 +69,8 @@ public abstract class ExpandedViewHolder extends ArrowItemViewHolder {
     @Override
     public Animator onAnimateChange(List<Object> payloads, int fromLeft, int fromTop, int fromRight,
                                     int fromBottom, long duration) {
-        if (payloads == null || payloads.isEmpty()) {
-            return null;
-        }
-
-        final AnimatorSet animatorSet = new AnimatorSet();
-        if (arrow != null) {
-            animatorSet.playTogether(AnimatorUtils.getBoundsAnimator(itemView,
-                    fromLeft, fromTop, fromRight, fromBottom,
-                    itemView.getLeft(), itemView.getTop(), itemView.getRight(), itemView.getBottom()),
-                    ObjectAnimator.ofFloat(arrow, TRANSLATION_Y, 0f));
-        } else {
-            animatorSet.playTogether(AnimatorUtils.getBoundsAnimator(itemView,
-                    fromLeft, fromTop, fromRight, fromBottom,
-                    itemView.getLeft(), itemView.getTop(), itemView.getRight(), itemView.getBottom()));
-        }
-        animatorSet.addListener(new AnimatorListenerAdapter() {
-            @Override
-            public void onAnimationEnd(Animator animator) {
-                setTranslationY(0f);
-                itemView.requestLayout();
-            }
-        });
-        animatorSet.setDuration(duration);
-        animatorSet.setInterpolator(AnimatorUtils.INTERPOLATOR_FAST_OUT_SLOW_IN);
-
-        return animatorSet;
-    }
-
-    private void setTranslationY(float translationY) {
-        if (arrow != null) {
-            arrow.setTranslationY(translationY);
-        }
+        /* There are no possible partial animations for collapsed view holders. */
+        return null;
     }
 
     @Override
@@ -124,17 +81,15 @@ public abstract class ExpandedViewHolder extends ArrowItemViewHolder {
             return null;
         }
 
-        final boolean isExpanding = this == newHolder;
-        AnimatorUtils.setBackgroundAlpha(itemView, isExpanding ? 0 : 255);
-        setChangingViewsAlpha(isExpanding ? 0f : 1f);
+        final boolean isCollapsing = this == newHolder;
+        setChangingViewsAlpha(isCollapsing ? 0f : 1f);
 
-        final Animator changeAnimatorSet = isExpanding
-                ? createExpandingAnimator((ArrowItemViewHolder) oldHolder, duration)
-                : createCollapsingAnimator((ArrowItemViewHolder) newHolder, duration);
+        final Animator changeAnimatorSet = isCollapsing
+                ? createCollapsingAnimator((ArrowItemViewHolder) oldHolder, duration)
+                : createExpandingAnimator((ArrowItemViewHolder) newHolder, duration);
         changeAnimatorSet.addListener(new AnimatorListenerAdapter() {
             @Override
             public void onAnimationEnd(Animator animator) {
-                AnimatorUtils.setBackgroundAlpha(itemView, 255);
                 if (arrow != null) {
                     arrow.setVisibility(View.VISIBLE);
                     arrow.setTranslationY(0f);
@@ -146,37 +101,28 @@ public abstract class ExpandedViewHolder extends ArrowItemViewHolder {
         return changeAnimatorSet;
     }
 
-    private Animator createCollapsingAnimator(ArrowItemViewHolder newHolder, long duration) {
+    private Animator createExpandingAnimator(ArrowItemViewHolder newHolder, long duration) {
         if (arrow != null) {
             arrow.setVisibility(View.INVISIBLE);
         }
 
         final View oldView = itemView;
         final View newView = newHolder.itemView;
-
-        final Animator backgroundAnimator = ObjectAnimator.ofPropertyValuesHolder(oldView,
-                PropertyValuesHolder.ofInt(AnimatorUtils.BACKGROUND_ALPHA, 255, 0));
-        backgroundAnimator.setDuration(duration);
-
-        final Animator boundsAnimator = AnimatorUtils.getBoundsAnimator(oldView, oldView, newView);
-        boundsAnimator.setDuration(duration);
+        final Animator boundsAnimator = AnimatorUtils.getBoundsAnimator(oldView, oldView, newView)
+                .setDuration(duration);
         boundsAnimator.setInterpolator(AnimatorUtils.INTERPOLATOR_FAST_OUT_SLOW_IN);
 
         final AnimatorSet animatorSet = new AnimatorSet();
-        animatorSet.playTogether(backgroundAnimator, boundsAnimator);
+        animatorSet.playTogether(boundsAnimator);
         return animatorSet;
     }
 
-    private Animator createExpandingAnimator(ArrowItemViewHolder oldHolder, long duration) {
+    private Animator createCollapsingAnimator(ArrowItemViewHolder oldHolder, long duration) {
         final View oldView = oldHolder.itemView;
         final View newView = itemView;
-        final Animator boundsAnimator = AnimatorUtils.getBoundsAnimator(newView, oldView, newView);
-        boundsAnimator.setDuration(duration);
+        final Animator boundsAnimator = AnimatorUtils.getBoundsAnimator(newView, oldView, newView)
+                .setDuration(duration);
         boundsAnimator.setInterpolator(AnimatorUtils.INTERPOLATOR_FAST_OUT_SLOW_IN);
-
-        final Animator backgroundAnimator = ObjectAnimator.ofPropertyValuesHolder(newView,
-                PropertyValuesHolder.ofInt(AnimatorUtils.BACKGROUND_ALPHA, 0, 255));
-        backgroundAnimator.setDuration(duration);
 
         final AnimatorSet animatorSet;
         if (arrow != null) {
@@ -186,7 +132,6 @@ public abstract class ExpandedViewHolder extends ArrowItemViewHolder {
             ((ViewGroup) newView).offsetDescendantRectToMyCoords(arrow, newArrowRect);
             ((ViewGroup) oldView).offsetDescendantRectToMyCoords(oldArrow, oldArrowRect);
             final float arrowTranslationY = oldArrowRect.bottom - newArrowRect.bottom;
-
             arrow.setTranslationY(arrowTranslationY);
             arrow.setVisibility(View.VISIBLE);
 
@@ -195,7 +140,7 @@ public abstract class ExpandedViewHolder extends ArrowItemViewHolder {
             arrowAnimation.setInterpolator(AnimatorUtils.INTERPOLATOR_FAST_OUT_SLOW_IN);
 
             animatorSet = new AnimatorSet();
-            animatorSet.playTogether(backgroundAnimator, boundsAnimator, arrowAnimation);
+            animatorSet.playTogether(boundsAnimator, arrowAnimation);
             animatorSet.addListener(new AnimatorListenerAdapter() {
                 @Override
                 public void onAnimationStart(Animator animator) {
@@ -204,7 +149,7 @@ public abstract class ExpandedViewHolder extends ArrowItemViewHolder {
             });
         } else {
             animatorSet = new AnimatorSet();
-            animatorSet.playTogether(backgroundAnimator, boundsAnimator);
+            animatorSet.playTogether(boundsAnimator);
         }
 
         return animatorSet;
