+++ fancyaccordionview/src/main/java/com/sysdata/widget/accordion/utils/AnimatorUtils.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.sysdata.widget.accordionview.utils;
+package com.sysdata.widget.accordion.utils;
 
 import android.animation.Animator;
 import android.animation.ObjectAnimator;
