+++ fancyaccordionview/src/test/java/com/sysdata/demo/accordionview/ExampleUnitTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.sysdata.widget.accordionview;
+package com.sysdata.widget.accordion;
 
 import org.junit.Test;
